pyproject.toml
[tool.poetry]
name = "scarcejealouspentagon"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
chatterbot-voice = "^0.1.5"
flask = "^1.1.2"
flask-restful = "^0.3.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
